[WIP] Update and simplify make scripts

+++ Makefile
@@ -1,13 +1,10 @@
 PWD=$(shell pwd)
 
 build:
-	./scripts/make.sh || true
+	./scripts/make.sh
 
 run:
 	DEBUG=1 SYSD_UI_DIR="$$(pwd)/mods/ui/files" sysd/sysd
-
-test:
-	env
 
 doc:
 	@echo "" > docs/source/modules.rst
@@ -15,9 +12,8 @@ doc:
 	cd docs; make html
 
 clean:
-	rm -rf .gopath || true
-	rm -rf .tmp || true
-	rm -rf sysd/sysd || true
+	[ -e .gopath ] && rm -rf .gopath || true
+	[ -e sysd/sysd ] && rm -rf sysd/sysd || true
 
 install:
 	install -D --mode=0644 sysd/sysd $(DESTDIR)/usr/sbin/sysd
